% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{covers}
\alias{covers}
\title{Check if a set covers a partition
check if a set covers a partition, that is: "it contains at least one element from each cell of the partition"}
\usage{
covers(set, partition)
}
\arguments{
\item{set}{a character vector containing members of the set}

\item{partition}{a list of character vector containing members of the partition}
}
\value{
A logical: TRUE if the set covers the partition and FALSE if not
}
\description{
Check if a set covers a partition
check if a set covers a partition, that is: "it contains at least one element from each cell of the partition"
}
