% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{data_example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 3 columns.
}
\usage{
data_example
}
\description{
A toy example of three factors with differents modalities, that can constitute a 3-partite graph
}
\keyword{datasets}
