% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{function_assign_node_to_partite}
\alias{function_assign_node_to_partite}
\title{Assign node to a set}
\usage{
function_assign_node_to_partite(node_names, dat)
}
\arguments{
\item{node_names}{the vector of node names of the graph}

\item{dat}{a datasets with multiple columns, each column containing each factor composing the k sets of the G}
}
\description{
Function to add each node to a given set
}
