% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{function_count_members_from_tibble}
\alias{function_count_members_from_tibble}
\title{Function to arrange rows by number of members of the kclique}
\usage{
function_count_members_from_tibble(
  tibble_k_clique,
  columns_to_select = everything(),
  sep = "-"
)
}
\arguments{
\item{tibble_k_clique}{output of function_list_kcliques_to_tibble function}

\item{columns_to_select}{Columns to for which the number of members has to be counted, can be a character vector or some selection helper (like everything() or last_col())}

\item{sep}{the separator used to separate kclique members in the input tibble}
}
\description{
Function to arrange rows by number of members of the kclique
}
