% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{function_detect_belonging}
\alias{function_detect_belonging}
\title{From a string containing the kcliques, detects to which set each node belongs}
\usage{
function_detect_belonging(string_input, dat, sep = ";")
}
\arguments{
\item{string_input}{input a character vector containing members of the kcliques}

\item{dat}{a dataset with multiple columns, each column containing each factor composing the k sets of the G}

\item{sep}{separator between node names}
}
\description{
From a string containing the kcliques, detects to which set each node belongs
}
