% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{get_adjacency_matrix_from_dat}
\alias{get_adjacency_matrix_from_dat}
\title{Compute adjacency matrix from matrix with factors as columns}
\usage{
get_adjacency_matrix_from_dat(dat)
}
\arguments{
\item{dat}{a dataset with multiple columns, each column containing each factor composing the k sets of the G}
}
\description{
Get the adjacency matrix linked to a table where each line represents a combination of factors
}
