% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{get_graph_from_dat}
\alias{get_graph_from_dat}
\title{Compute graph from matrix with factors as columns}
\usage{
get_graph_from_dat(dat)
}
\arguments{
\item{dat}{a dataset with multiple columns, each column containing each factor composing the k sets of the G}
}
\value{
a k-partite graph, from class graph (package igraph).
}
\description{
Get the graph linked to a table where each line represents a combination of factors
}
