% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{maximal_kclique_enumeration}
\alias{maximal_kclique_enumeration}
\title{Maximal multipartite clique enumeration}
\usage{
maximal_kclique_enumeration(G, dat)
}
\arguments{
\item{G}{a k-partite graph, from class graph (package igraph).}

\item{dat}{a dataset with multiple columns, each column containing each factor composing the k sets of the G}
}
\value{
A character vector containing all maximal kcliques ; each element of the vector is a maximal kclique ; each member of each clique is separated by a ";"
}
\description{
Function to enumerate all maximal kcliques within a k-partite graph. It is a direct implementation of
the algorithms used come from this paper Phillips et al. \url{doi:10.3390/a12010023}.
}
