% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot multipartite graph}
\usage{
plot_graph(
  G,
  dat,
  layout_name = "in_circle",
  size_text = 3.5,
  color_text = "black",
  color_edges = "black",
  alpha_node = 1
)
}
\arguments{
\item{G}{a k-partite graph, from class graph (package igraph).}

\item{dat}{a datasets with multiple columns, each column containing each factor composing the k sets of the G}

\item{layout_name}{The type of layout to create. Either a valid string, a function, a matrix, or a data.frame (see help(ggraph::ggraph))}

\item{size_text}{a numeric passed to the size argument of function geom_node_text, default 3.5}

\item{color_text}{a character passed to the colour argument of function geom_node_text, default "black"}

\item{color_edges}{a character passed to the colour argument of function geom_edge_link, default "black"}

\item{alpha_node}{a numeric between 0 and 1 giving the level of transparency wanted for the nodes, default 1}
}
\value{
An object of class ggraph::gg onto which layers, scales,
etc. can be added.
}
\description{
Function to plot a multipartite graph with a given layout
}
