% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_kclique}
\alias{plot_kclique}
\title{Plot a kclique}
\usage{
plot_kclique(
  G,
  tibble_kcliques,
  dataset,
  vector_kcliques = 1,
  automatic_nodes = TRUE,
  show_legend_edges = FALSE,
  color_edge_kclique = "black",
  color_edge_no_kclique = "grey75",
  layout_name = "in_circle"
)
}
\arguments{
\item{G}{a k-partite graph, from class graph (package igraph).}

\item{tibble_kcliques}{An output of the function function_list_kcliques_to_tibble}

\item{dataset}{a datasets with multiple columns, each column containing each factor composing the k sets of the G}

\item{vector_kcliques}{a vector containing the rows of the tibble containing the kcliques you want to highlight}

\item{automatic_nodes}{a logical telling if the nodes have to be colored accordingly to their partite's belonging}

\item{show_legend_edges}{a logical telling if the edges' color legend has to be shown}

\item{color_edge_kclique}{a string indicating the color of the edges members of the kclique}

\item{color_edge_no_kclique}{a string indicating the color of the edges that are not members of the kclique}

\item{layout_name}{The type of layout to create. Either a valid string, a function, a matrix, or a data.frame (see help(ggraph::ggraph))}
}
\value{
An object of class ggraph::gg onto which layers, scales,
etc. can be added.
}
\description{
Function to highlight a kclique one a graph
}
