% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate the cliques of a graph}
\usage{
enumerate(G, R, P, X, partition, cliques_discovered)
}
\arguments{
\item{G}{a k-partite graph, from class graph (package igraph).}

\item{R}{character vector containing the current clique}

\item{P}{character vector containing the vertices that can extend the current clique}

\item{X}{character vector containing the vertices thathave already been considered}

\item{partition}{a list of character vector containing members of the partition}

\item{cliques_discovered}{a list of the previously discovered kcliques}
}
\value{
A logical: TRUE if the set covers the partition and FALSE if not
}
\description{
Sub Routine enumerate described in \url{doi:10.3390/a12010023}.
}
