% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kcliques.R
\name{function_list_kcliques_to_tibble}
\alias{function_list_kcliques_to_tibble}
\title{Return formatted data from a kclique object}
\usage{
function_list_kcliques_to_tibble(kcliques, dat)
}
\arguments{
\item{kcliques}{output of maximal_kclique_enumeration function}

\item{dat}{a dataset with multiple columns, each column containing each factor composing the k sets of the G}
}
\description{
Return formatted data from a kclique object
}
