% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot multipartite graph}
\usage{
plot_graph(G, dat, layout_name = "in_circle")
}
\arguments{
\item{G}{a k-partite graph, from class graph (package igraph).}

\item{dat}{a datasets with multiple columns, each column containing each factor composing the k sets of the G}

\item{layout_name}{The type of layout to create. Either a valid string, a function, a matrix, or a data.frame (see help(ggraph::ggraph))}
}
\value{
An object of class ggraph::gg onto which layers, scales,
etc. can be added.
}
\description{
Function to plot a multipartite graph with a given layout
}
